Created: - AddVehicleView FXML - addded electric type - AddVehicleViewController - edited - AddVehicleViewModel - edited - Vehicle - edited getPrice()

+++ src/Client/Model/Vehicle.java
@@ -107,7 +107,7 @@ public class Vehicle
     this.price = price;
   }
 
-  public int getPrice()
+  public double getPrice()
   {
     return price;
   }
